Extract generic `AuthOnlyDjangoFormMutation`
