+++ backend/api/mutations.py
@@ -1,3 +1,5 @@
+from graphql import GraphQLError
+
 from graphene_django.forms import mutation
 
 
@@ -10,3 +12,15 @@ class ContextAwareDjangoModelFormMutation(mutation.DjangoModelFormMutation):
         kwargs = super().get_form_kwargs(root, info, **input)
         kwargs['context'] = info.context
         return kwargs
+
+
+class AuthOnlyDjangoFormMutation(ContextAwareDjangoModelFormMutation):
+    class Meta:
+        abstract = True
+
+    @classmethod
+    def mutate_and_get_payload(cls, root, info, **input):
+        if not info.context.user.is_authenticated:
+            raise GraphQLError('User not logged in')
+
+        return super().mutate_and_get_payload(root, info, **input)
